+++ src/applications/pre-need-integration/config/pages/burialBenefits.jsx
@@ -1,33 +1,11 @@
 import fullSchemaPreNeed from 'vets-json-schema/dist/40-10007-schema.json';
 import React from 'react';
 
-import * as autosuggest from 'platform/forms-system/src/js/definitions/autosuggest';
-
-import { useSelector } from 'react-redux';
-import {
-  isVeteran,
-  getCemeteries,
-  desiredCemeteryNoteDescriptionVeteran,
-  desiredCemeteryNoteDescriptionNonVeteran,
-} from '../../utils/helpers';
+import { isVeteran } from '../../utils/helpers';
 
 const {
   hasCurrentlyBuried,
 } = fullSchemaPreNeed.properties.application.properties;
-
-function DesiredCemeteryNoteDescription() {
-  const data = useSelector(state => state.form.data || {});
-  return isVeteran(data)
-    ? desiredCemeteryNoteDescriptionVeteran
-    : desiredCemeteryNoteDescriptionNonVeteran;
-}
-
-function DesiredCemeteryTitle() {
-  const data = useSelector(state => state.form.data || {});
-  return isVeteran(data)
-    ? 'Which VA national cemetery would you prefer to be buried in?'
-    : 'Which VA national cemetery would the applicant prefer to be buried in?';
-}
 
 export const desiredCemeteryNoteTitleWrapper = (
   <a
@@ -40,31 +18,8 @@ export const desiredCemeteryNoteTitleWrapper = (
   </a>
 );
 
-function DesiredCemeteryNoteTitle() {
-  return desiredCemeteryNoteTitleWrapper;
-}
-
-export const desiredCemeteryTitleWrapper = <DesiredCemeteryTitle />;
-
 export const uiSchema = {
   application: {
-    claimant: {
-      desiredCemetery: autosuggest.uiSchema(
-        desiredCemeteryTitleWrapper,
-        getCemeteries,
-        {
-          'ui:options': {
-            inputProps: {
-              'aria-describedby': 'burial-cemetary-note',
-            },
-          },
-        },
-      ),
-      'view:desiredCemeteryNote': {
-        'ui:title': DesiredCemeteryNoteTitle,
-        'ui:description': DesiredCemeteryNoteDescription,
-      },
-    },
     hasCurrentlyBuried: {
       'ui:widget': 'radio',
       'ui:options': {
@@ -95,16 +50,6 @@ export const schema = {
       type: 'object',
       required: ['hasCurrentlyBuried'],
       properties: {
-        claimant: {
-          type: 'object',
-          properties: {
-            desiredCemetery: autosuggest.schema,
-            'view:desiredCemeteryNote': {
-              type: 'object',
-              properties: {},
-            },
-          },
-        },
         hasCurrentlyBuried,
       },
     },
